// web/src/api/models/EventSummary.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { i64 } from "./i64";
/**
 * Summary event response for list views
 */
export type EventSummary = {
  /**
   * Creation timestamp
   */
  created: string;
  /**
   * Whether event has payload data
   */
  has_payload: boolean;
  /**
   * Event ID
   */
  id: i64;
  /**
   * Event payload data, when present
   */
  payload: any | null;
  rule?: null | i64;
  /**
   * Rule reference (if event was generated by a specific rule)
   */
  rule_ref?: string | null;
  source?: null | i64;
  /**
   * Source reference
   */
  source_ref?: string | null;
  /**
   * Trace tag associated to this event.
   */
  trace_tag?: string | null;
  trigger?: null | i64;
  /**
   * Trigger reference
   */
  trigger_ref: string;
};